MissionTie hook

+++ XwaHookDialog.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
+﻿using Idmr.Platform.Xwa;
+using System;
 using System.IO;
 using System.Windows.Forms;
 
@@ -29,17 +24,25 @@ namespace Idmr.Yogeme
 		string _res = "\\Resdata\\";
 		string _wave = "\\Wave\\";
 		string _fm = "\\FlightModels\\";
+		enum ReadMode { None = -1, Backdrop, Mission }
 
-		public XwaHookDialog(string missionFile)
+		public XwaHookDialog(Mission mission)
 		{
 			InitializeComponent();
-			_mission = Idmr.Common.StringFunctions.GetFileName(missionFile, false);
+			_mission = Idmr.Common.StringFunctions.GetFileName(mission.MissionPath, false);
 			if (_mission == "NewMission")
 			{
 				MessageBox.Show("Please perform inital save prior to hook assignment.", "New Mission detected", MessageBoxButtons.OK, MessageBoxIcon.Warning);
 				cmdCancel_Click("NewMission", new EventArgs());
 			}
-			_fileName = missionFile.Replace(".tie", ".ini");
+			_fileName = mission.MissionPath.Replace(".tie", ".ini");
+
+			cboIff.Items.AddRange(Strings.IFF);
+			for (int i = cboIff.Items.Count; i < 256; i++) cboIff.Items.Add("IFF #" + (i + 1));
+			cboMarkings.Items.AddRange(Strings.Color);
+			for (int i = cboMarkings.Items.Count; i < 256; i++) cboMarkings.Items.Add("Clr #" + (i + 1));
+			cboFG.Items.AddRange(mission.FlightGroups.GetList());
+
 			Settings s = new Settings();
 			if (s.XwaInstalled)
 			{
@@ -53,13 +56,15 @@ namespace Idmr.Yogeme
 				_hangarObjectsFile = checkFile("_HangarObjects.txt");
 				_hangarCameraFile = checkFile("_HangarCamera.txt");
 				_famHangarCameraFile = checkFile("_FamHangarCamera.txt");
-				_hangarMapFile = checkFile("_HamgarMap.txt");
+				_hangarMapFile = checkFile("_HangarMap.txt");
 				_famHangarMapFile = checkFile("_FamHangarMap.txt");
 			}
 			StreamReader srMission = null;
 			string line = "";
-			bool readLine = false;
 			if (File.Exists(_fileName)) srMission = new StreamReader(_fileName);
+			ReadMode readMode = ReadMode.None;
+
+			#region individual files
 			if (_bdFile != "")
 			{
 				StreamReader srBD = new StreamReader(_bdFile);
@@ -69,19 +74,62 @@ namespace Idmr.Yogeme
 				}
 				srBD.Close();
 			}
-			else if (srMission != null)
+			if (_missionTxtFile != "")
+			{
+				StreamReader srMiss = new StreamReader(_missionTxtFile);
+				while((line = srMiss.ReadLine()) != null)
+				{
+					line = line.ToLower().Replace(" ", "");
+					string[] parts = line.Split(',');
+					if (parts.Length == 4 && parts[0] == "fg")
+					{
+						int fg = int.Parse(parts[1]);
+						if (parts[2] == "markings")
+							lstMission.Items.Add(fg + "," + cboFG.Items[fg].ToString() + ",marks," + cboMarkings.Items[int.Parse(parts[3])].ToString());
+						else if (parts[2] == "iff")
+							lstMission.Items.Add(fg + "," + cboFG.Items[fg].ToString() + ",iff," + cboIff.Items[int.Parse(parts[3])].ToString());
+						else if (parts[2] == "pilotvoice")
+							lstMission.Items.Add(fg + "," + cboFG.Items[fg].ToString() + ",pilot," + parts[3]);
+					}
+				}
+				srMiss.Close();
+			}
+			#endregion
+
+			if (srMission != null)
 			{
 				while ((line = srMission.ReadLine()) != null)
 				{
-					// TODO: this will need to be redone so it'll read in any order
-					if (line.StartsWith("[")) readLine = false;
-					if (readLine) lstBackdrops.Items.Add(line);
-					else if (line.ToLower() == "[resdata]") readLine = true;
+					if (line == "" || line.Trim().StartsWith(";")) continue;
+
+					if (line.StartsWith("["))
+					{
+						readMode = ReadMode.None;
+						if (line.ToLower() == "[resdata]") readMode = ReadMode.Backdrop;
+						else if (line.ToLower() == "[mission_tie]") readMode = ReadMode.Mission;
+					}
+					else if (readMode == ReadMode.Backdrop) lstBackdrops.Items.Add(line);
+					else if (readMode == ReadMode.Mission)
+					{
+						line = line.ToLower().Replace(" ", "");
+						string[] parts = line.Split(',');
+						if (parts.Length == 4 && parts[0] == "fg")
+						{
+							int fg = int.Parse(parts[1]);
+							if (parts[2] == "markings")
+								lstMission.Items.Add(cboFG.Items[fg].ToString() + ",marks," + cboMarkings.Items[int.Parse(parts[3])].ToString());
+							else if (parts[2] == "iff")
+								lstMission.Items.Add(cboFG.Items[fg].ToString() + ",iff," + cboIff.Items[int.Parse(parts[3])].ToString());
+							else if (parts[2] == "pilotvoice")
+								lstMission.Items.Add(cboFG.Items[fg].ToString() + ",pilot," + parts[3]);
+						}
+					}
 				}
 			}
 
 			srMission.Close();
 			chkBackdrops.Checked = (lstBackdrops.Items.Count > 0);
+			chkMission.Checked = (lstMission.Items.Count > 0);
 		}
 
 		string checkFile(string extension)
@@ -114,16 +162,51 @@ namespace Idmr.Yogeme
 		}
 		#endregion Backdrops
 
+		#region MissionTie
+		private void chkMission_CheckedChanged(object sender, EventArgs e)
+		{
+			lstMission.Enabled = chkMission.Checked;
+			cmdAddMiss.Enabled = chkMission.Checked;
+			cmdRemoveMiss.Enabled = chkMission.Checked;
+			cboFG.Enabled = chkMission.Checked;
+			optMarkings.Enabled = chkMission.Checked;
+			optIff.Enabled = chkMission.Checked;
+			optPilot.Enabled = chkMission.Checked;
+			cboMarkings.Enabled = chkMission.Checked;
+			cboIff.Enabled = chkMission.Checked;
+			txtPilot.Enabled = chkMission.Checked;
+		}
+
+		private void cmdAddMiss_Click(object sender, EventArgs e)
+		{
+			if (cboFG.SelectedIndex == -1 || (optMarkings.Checked && cboMarkings.SelectedIndex == -1) || (optIff.Checked && cboIff.SelectedIndex == -1)
+				|| (optPilot.Checked && txtPilot.Text == "")) return;
+
+			if (optMarkings.Checked) lstMission.Items.Add(cboFG.Text + ",marks," + cboMarkings.Text);
+			else if (optIff.Checked) lstMission.Items.Add(cboFG.Text + ",iff," + cboIff.Text);
+			else if (optPilot.Checked) lstMission.Items.Add(cboFG.Text + ",pilot," + txtPilot.Text);
+
+		}
+		private void cmdRemoveMiss_Click(object sender, EventArgs e)
+		{
+			if (lstMission.SelectedIndex == -1) return;
+			lstMission.Items.RemoveAt(lstMission.SelectedIndex);
+		}
+		#endregion
+
 		private void cmdCancel_Click(object sender, EventArgs e)
 		{
 			Close();
 		}
 		private void cmdOK_Click(object sender, EventArgs e)
 		{
-			if (!chkBackdrops.Checked)
+			if (!chkBackdrops.Checked && _bdFile != "") File.Delete(_bdFile);
+
+			if (!chkMission.Checked && _missionTxtFile != "") File.Delete(_missionTxtFile);
+
+			if (!chkBackdrops.Checked && !chkMission.Checked)
 			{
 				File.Delete(_fileName);
-				if (_bdFile != "") File.Delete(_bdFile);
 				Close();
 				return;
 			}
@@ -147,19 +230,52 @@ namespace Idmr.Yogeme
 					for (int i = 0; i < lstBackdrops.Items.Count; i++) sw.WriteLine(lstBackdrops.Items[i]);
 					sw.WriteLine("");
 				}
+				if (chkMission.Checked && lstMission.Items.Count > 0)
+				{
+					sw.WriteLine("[Mission_Tie]");
+					for(int i = 0; i < lstMission.Items.Count; i++)
+					{
+						string[] parts = lstMission.Items[i].ToString().Split(',');
+						int fg;
+						for (fg = 0; fg < cboFG.Items.Count; fg++) if (cboFG.Items[fg].ToString() == parts[0]) break;
+						if (parts[1] == "marks")
+						{
+							for (int m = 0; m < cboMarkings.Items.Count; m++)
+								if (cboMarkings.Items[m].ToString() == parts[2])
+								{
+									sw.WriteLine("fg, " + fg + ", markings, " + m);
+									break;
+								}
+						}
+						else if (parts[1] == "iff")
+						{
+							for (int iff = 0; iff < cboIff.Items.Count; iff++)
+								if (cboIff.Items[iff].ToString() == parts[2])
+								{
+									sw.WriteLine("fg, " + fg + ", iff, " + iff);
+									break;
+								}
+						}
+						else if (parts[1] == "pilot")
+							sw.WriteLine("fg, " + fg + ", pilotvoice, " + parts[2]);
+					}
+					sw.WriteLine("");
+				}
 				sw.Flush();
 				sw.Close();
 				if (_bdFile != "") File.Delete(_bdFile);
+				if (_missionTxtFile != "") File.Delete(_missionTxtFile);
 			}
 			catch
 			{
 				if (sw != null) sw.Close();
 				if (File.Exists(backup))
 				{
+					File.Delete(_fileName);
 					File.Copy(backup, _fileName);
-					File.Delete(backup);
 				}
 			}
+			File.Delete(backup);
 			Close();
 		}
 	}
